fix: move socialIcons and skills to client component
Resolve prerender error by moving icon components and data from server
component (page.tsx) to HeroWrapper client component boundary.
