+++ app/components/sections/herowrapper.tsx
@@ -1,14 +1,41 @@
 "use client";
 
-import { useState, useEffect } from 'react';
-import Hero from './hero';
+import { useState, useEffect } from "react";
+import { Github, Twitter, Linkedin, Mail } from 'lucide-react';
+import { SiBluesky } from '@icons-pack/react-simple-icons';
+import Hero from "./hero";
 
-interface HeroWrapperProps {
-  skills: string[];
-  socialIcons: any[];
-}
+const socialIcons = [
+  {
+    icon: Github,
+    href: "https://github.com/machage9603",
+    label: "GitHub",
+  },
+  {
+    icon: Twitter,
+    href: "https://twitter.com/sermachage",
+    label: "Twitter",
+  },
+  {
+    icon: SiBluesky,
+    href: "https://bsky.app/profile/sermachage.bsky.social",
+    label: "Bluesky",
+  },
+  {
+    icon: Linkedin,
+    href: "https://www.linkedin.com/in/mikemachage/",
+    label: "LinkedIn",
+  },
+  {
+    icon: Mail,
+    href: "mailto:machage@example.com",
+    label: "Email",
+  },
+];
 
-export default function HeroWrapper({ skills, socialIcons }: HeroWrapperProps) {
+const skills = ["Golang", "Rust", "TypeScript"];
+
+export default function HeroWrapper() {
   const [isVisible, setIsVisible] = useState(false);
   const [currentSkill, setCurrentSkill] = useState(0);
 
@@ -16,16 +43,16 @@ export default function HeroWrapper({ skills, socialIcons }: HeroWrapperProps) {
     setIsVisible(true);
     const interval = setInterval(() => {
       setCurrentSkill((prev) => (prev + 1) % skills.length);
-    }, 3000);
+    }, 2000);
     return () => clearInterval(interval);
-  }, [skills.length]);
+  }, []);
 
   return (
-    <Hero 
-      isVisible={isVisible} 
-      skills={skills} 
-      currentSkill={currentSkill} 
-      socialIcons={socialIcons} 
+    <Hero
+      isVisible={isVisible}
+      skills={skills}
+      currentSkill={currentSkill}
+      socialIcons={socialIcons}
     />
   );
 }
