+++ public/sitemap-0.xml
@@ -1,8 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>http://mikemachage.tech/projects</loc><lastmod>2025-10-01T20:03:21.227Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>http://mikemachage.tech/about</loc><lastmod>2025-10-01T20:03:21.227Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>http://mikemachage.tech/contact</loc><lastmod>2025-10-01T20:03:21.227Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>http://mikemachage.tech</loc><lastmod>2025-10-01T20:03:21.227Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>http://mikemachage.tech/blog</loc><lastmod>2025-10-01T20:03:21.227Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>http://mikemachage.tech</loc><lastmod>2025-10-06T22:58:30.625Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
 </urlset>
